+++ src/Preview/Preview.tsx
@@ -14,7 +14,7 @@
  * ======================================================================== */
 import * as React from 'react';
 import {
-  Grid, SelectChangeEvent,
+  Theme, Grid, SelectChangeEvent,
 } from '@mui/material';
 import { styled } from '@mui/material/styles';
 import IconDownload from '@hcl-software/enchanted-icons/dist/carbon/es/download';
@@ -53,6 +53,7 @@ import { white } from '../colors';
  * @property {function} overrideHandleNext - Event handler that overrides the default next asset navigation
  * @property {boolean} isFetchingAssets - indicate whether the assets prop is still fetching, Note: even when this is false and asset is not finished rendering, <Preview /> will still show spinner
  * @property {boolean} customHeaderTitle - custom header title
+ * @property {boolean} isVersionComparison - check if the preivew is used for version comparison
 */
 
 export interface MediaType {
@@ -98,6 +99,7 @@ export interface PreviewProps {
   isFetchingAssets?: boolean;
   customHeaderTitle?: string;
   handleError?: (event: React.SyntheticEvent<HTMLVideoElement | HTMLImageElement, Event>) => void;
+  isVersionComparison?: boolean;
 }
 
 // Zoom button margin is 12px
@@ -294,6 +296,7 @@ const Preview: React.FC<PreviewProps> = ({
   isFetchingAssets = false,
   customHeaderTitle,
   handleError,
+  isVersionComparison = false,
 }: PreviewProps) => {
   const fallbackAssetValue: Assets[] = [
     {
@@ -607,6 +610,7 @@ const Preview: React.FC<PreviewProps> = ({
             sx={{
               transform: `scale(${zoomPercentage / 100})`,
               opacity: isCurrentAssetReady ? 1 : 0,
+              transformOrigin: 'center',
             }}
             draggable="true"
             onLoad={handleImageLoad}
@@ -623,15 +627,23 @@ const Preview: React.FC<PreviewProps> = ({
       open={open}
       sx={{
         zIndex: 999,
+        position: isVersionComparison ? 'relative' : 'fixed',
+        backgroundColor: (scopedTheme: Theme) => { return scopedTheme.palette.background.secondary; },
       }}
     >
       <PreviewContainer
         container
         direction="column"
         justifyContent="flex-start"
+        sx={{
+          height: isVersionComparison ? '400px' : '100vh',
+        }}
       >
         <Grid>
           <Header
+            sx={{
+              display: isVersionComparison ? 'none' : '',
+            }}
             startSection={{
               hamburgerSpace: false,
               withBackButton: true,
@@ -669,17 +681,20 @@ const Preview: React.FC<PreviewProps> = ({
                 placement="bottom"
                 title={tooltipTexts.download}
               >
-                <IconButton
-                  data-testid={PreviewTestIds.PREVIEW_DOWNLOAD_BUTTON}
-                  variant={IconButtonVariants.WITH_PADDING}
-                  disabled={!isCurrentAssetReady || reactComponent !== undefined}
-                  onClick={(e) => {
-                    const selectRenditionId = currentRendition.id;
-                    if (handleDownload) handleDownload(e, selectRenditionId);
-                  }}
-                >
-                  <IconDownload />
-                </IconButton>
+                <span>
+                  <IconButton
+                    data-testid={PreviewTestIds.PREVIEW_DOWNLOAD_BUTTON}
+                    variant={IconButtonVariants.WITH_PADDING}
+                    disabled={!isCurrentAssetReady || reactComponent !== undefined || isVersionComparison}
+                    onClick={(e) => {
+                      const selectRenditionId = currentRendition.id;
+                      if (handleDownload) handleDownload(e, selectRenditionId);
+                    }}
+                    showendicon={0}
+                  >
+                    <IconDownload />
+                  </IconButton>
+                </span>
               </Tooltip>,
               <Button
                 data-testid={PreviewTestIds.PREVIEW_SELECT_BUTTON}
@@ -704,60 +719,75 @@ const Preview: React.FC<PreviewProps> = ({
             justifyContent="space-between"
             alignItems="center"
             sx={{
-              height: `calc(100% - ${isVideo ? '55' : '61'}px)`,
+              height: `calc(100% - ${isVideo ? '55' : `${isVersionComparison ? '0' : '66'}`}px)`,
+              overflow: (zoomPercentage === zoomDefault || zoomPercentage === zoomToFitPercentage) ? 'hidden' : 'auto',
             }}
           >
-            <Tooltip
-              tooltipsize="small"
-              placement="bottom-start"
-              title={tooltipTexts.previousAsset}
-            >
-              <PreviousPreviewButton>
-                <StyledArrowButton
-                  data-testid={PreviewTestIds.PREVIEW_PREV_BUTTON}
-                  disabled={isPreviousDisabled}
-                  onClick={handlePreviousAsset}
-                >
-                  <ChevronLeft />
-                </StyledArrowButton>
-              </PreviousPreviewButton>
-            </Tooltip>
+            { !isVersionComparison && (
+              <Tooltip
+                tooltipsize="small"
+                placement="bottom-start"
+                title={tooltipTexts.previousAsset}
+              >
+                <PreviousPreviewButton>
+                  <span>
+                    <StyledArrowButton
+                      data-testid={PreviewTestIds.PREVIEW_PREV_BUTTON}
+                      disabled={isPreviousDisabled || isVersionComparison}
+                      onClick={handlePreviousAsset}
+                      showendicon={0}
+                    >
+                      <ChevronLeft />
+                    </StyledArrowButton>
+                  </span>
+                </PreviousPreviewButton>
+              </Tooltip>
+            )}
             <>
               {(isCurrentAssetReady === false && reactComponent === undefined) && (
               <CircularProgressContainer
                 data-testid={PreviewTestIds.PREVIEW_CIRCULAR_PROGRESS}
                 container
                 sx={{
-                  height: `calc(100% - ${isVideo ? '55' : '61'}px)`,
+                  height: `calc(100% - ${isVideo ? '55' : `${isVersionComparison ? '0' : '66'}`}px)`,
                 }}
               >
-                <CircularProgress />
+                <CircularProgress withbackdrop={0} />
               </CircularProgressContainer>
               )}
               {renderOptions()}
             </>
-            <Tooltip
-              tooltipsize="small"
-              placement="bottom-end"
-              title={tooltipTexts.nextAsset}
-            >
-              <NextPreviewButton>
-                <StyledArrowButton
-                  data-testid={PreviewTestIds.PREVIEW_NEXT_BUTTON}
-                  disabled={isNextDisabled}
-                  onClick={handleNextAsset}
-                >
-                  <ChevronRight />
-                </StyledArrowButton>
-              </NextPreviewButton>
-            </Tooltip>
+            { !isVersionComparison && (
+              <Tooltip
+                tooltipsize="small"
+                placement="bottom-end"
+                title={tooltipTexts.nextAsset}
+              >
+                <NextPreviewButton>
+                  <span>
+                    <StyledArrowButton
+                      data-testid={PreviewTestIds.PREVIEW_NEXT_BUTTON}
+                      disabled={isNextDisabled || isVersionComparison}
+                      onClick={handleNextAsset}
+                      showendicon={0}
+                    >
+                      <ChevronRight />
+                    </StyledArrowButton>
+                  </span>
+                </NextPreviewButton>
+              </Tooltip>
+            )}
           </ImageContainer>
           {(!isVideo && isCurrentAssetReady && reactComponent === undefined) && (
           <Grid
             container
             justifyContent="center"
           >
-            <ZoomContainer>
+            <ZoomContainer
+              sx={{
+                position: isVersionComparison ? 'absolute' : 'fixed',
+              }}
+            >
               <Tooltip
                 tooltipsize="small"
                 placement="top"
@@ -768,6 +798,7 @@ const Preview: React.FC<PreviewProps> = ({
                   variant={IconButtonVariants.WITHOUT_PADDING}
                   disabled={zoomOutDisable}
                   onClick={handleZoomOut}
+                  showendicon={0}
                 >
                   <IconZoomOut />
                 </StyledIconButton>
@@ -797,6 +828,7 @@ const Preview: React.FC<PreviewProps> = ({
                   variant={IconButtonVariants.WITHOUT_PADDING}
                   disabled={zoomInDisable}
                   onClick={handleZoomIn}
+                  showendicon={0}
                 >
                   <IconZoomIn />
                 </StyledIconButton>