+++ src/PreviewAccordion/PreviewAccordionDetails.tsx
@@ -12,23 +12,28 @@
  * See the License for the specific language governing permissions and      *
  * limitations under the License.                                           *
  * ======================================================================== */
-import { ColorType } from './index';
 
-const hclsoftwareblue: ColorType = {
-  HCLSOFTWAREBLUE01: '#000032',
-  HCLSOFTWAREBLUE02: '#00005A',
-  HCLSOFTWAREBLUE03: '#000078',
-  HCLSOFTWAREBLUE04: '#0000AA',
-  HCLSOFTWAREBLUE05: '#001EC8',
-  HCLSOFTWAREBLUE06: '#003CE6',
-  HCLSOFTWAREBLUE07: '#0550DC',
-  HCLSOFTWAREBLUE07_12: 'rgba(5, 80, 220, .12)',
-  HCLSOFTWAREBLUE07_8: 'rgba(5, 80, 220, .08)',
-  HCLSOFTWAREBLUE07_20: 'rgba(5, 80, 220, .2)',
-  HCLSOFTWAREBLUE08: '#0F5FDC',
-  HCLSOFTWAREBLUE09: '#B3D9F8',
-  HCLSOFTWAREBLUE09_12: 'rgba(179, 217, 248, .12)',
-  HCLSOFTWAREBLUE09_8: 'rgba(179, 217, 248, .08)',
+import React from 'react';
+import { styled } from '@mui/material/styles';
+import AccordionDetails, { AccordionDetailsProps } from '../Accordion/AccordionDetails';
+
+export interface PreviewAccordionDetailsProps extends AccordionDetailsProps {
+}
+
+const StyledAccordionDetails = styled(AccordionDetails)(() => {
+  return {
+    '&:last-of-type': {
+      borderBottomLeftRadius: '4px',
+      borderBottomRightRadius: '4px',
+    },
+  };
+});
+
+const PreviewAccordionDetails = ({ ...props }: PreviewAccordionDetailsProps) => {
+  return <StyledAccordionDetails {...props} />;
 };
 
-export default hclsoftwareblue;
+PreviewAccordionDetails.defaultProps = {
+};
+
+export default PreviewAccordionDetails;